build: create lib and std modules
