+++ core/nodes_test.ts
@@ -1,6 +1,6 @@
 import { assertEquals } from "@std/assert";
 import { afterAll, beforeAll, describe, it } from "@std/testing/bdd";
-import { MockWebSocket } from "../lib/testing.ts";
+import { MockWebSocket } from "@lophus/lib/testing";
 import { NostrNode, NostrNodeBase } from "./nodes.ts";
 
 describe("NostrNodeBase", () => {
