+++ deno.json
@@ -5,6 +5,18 @@
   "imports": {
     "@noble/curves": "npm:@noble/curves@^1.3.0",
     "@noble/hashes": "npm:@noble/hashes@^1.3.3",
+    "@lophus/core/clients": "./core/clients.ts",
+    "@lophus/core/protocol": "./core/protocol.ts",
+    "@lophus/core/relays": "./core/relays.ts",
+    "@lophus/lib/types": "./lib/types.ts",
+    "@lophus/lib/testing": "./lib/testing.ts",
+    "@lophus/lib/streams": "./lib/streams.ts",
+    "@lophus/std/env": "./std/env.ts",
+    "@lophus/std/events": "./std/events.ts",
+    "@lophus/std/notes": "./std/notes.ts",
+    "@lophus/std/pools": "./std/pools.ts",
+    "@lophus/std/signs": "./std/signs.ts",
+    "@lophus/std/times": "./std/times.ts",
     "@std/assert": "jsr:@std/assert@^0.219.1",
     "@std/dotenv": "jsr:@std/dotenv@^0.219.1",
     "@std/streams": "jsr:@std/streams@^0.219.1",
@@ -21,6 +33,8 @@
     "update:commit": "deno task -q update --commit --prefix 'build(deps):' --prefix-lock 'build(lock):'"
   },
   "workspaces": [
-    "./core"
+    "./core",
+    "./lib",
+    "./std"
   ]
 }
