+++ deno.lock
@@ -125,7 +125,8 @@
       "npm:@noble/hashes@^1.3.3"
     ],
     "members": {
-      "@lophus/core": {}
+      "@lophus/core": {},
+      "@lophus/lib": {}
     }
   }
 }
