+++ lib/streams_test.ts
@@ -1,6 +1,6 @@
-import { assertArrayIncludes, assertEquals } from "@std/assert";
-import { describe, it } from "@std/testing/bdd";
-import { Distinctor } from "./streams.ts";
+import { assert, assertArrayIncludes, assertEquals } from "@std/assert";
+import { beforeEach, describe, it } from "@std/testing/bdd";
+import { ConsoleLogger, Distinctor } from "@lophus/lib/streams";
 
 describe("Distinctor", () => {
   it("filters out duplicate values from a stream", async () => {
@@ -18,5 +18,30 @@ describe("Distinctor", () => {
     );
     assertEquals(values.length, 3);
     assertArrayIncludes(values, [1, 2, 3]);
+  });
+});
+
+describe("ConsoleLogger", () => {
+  let output = "";
+
+  beforeEach(() => {
+    globalThis.console = new Proxy(globalThis.console, {
+      get: () => {
+        return (data: unknown) => {
+          output += data;
+        };
+      },
+    });
+  });
+
+  it("should be a writable stream", () => {
+    const logger = new ConsoleLogger();
+    assert(logger instanceof WritableStream);
+  });
+
+  it("should log to console", async () => {
+    const logger = new ConsoleLogger();
+    await logger.getWriter().write("ConsoleLogger");
+    assertEquals(output, "ConsoleLogger");
   });
 });
