+++ std/signs_test.ts
@@ -1,8 +1,7 @@
-import { describe, it } from "@std/testing/bdd";
 import { assert, assertEquals } from "@std/assert";
-import { Stringified } from "../lib/types.ts";
-import "../nips/01/protocol.ts";
-import { Timestamp } from "./times.ts";
+import { describe, it } from "@std/testing/bdd";
+import { Stringified } from "@lophus/lib/types";
+import { Timestamp } from "@lophus/std/times";
 import {
   PrivateKey,
   PublicKey,
