+++ std/times.ts
@@ -1,8 +1,12 @@
-import { Brand } from "../lib/types.ts";
+import { Brand } from "@lophus/lib/types";
 
 export type Timestamp = Brand<number, "EventTimeStamp">;
 
-export const Timestamp = {
+export type TimestampConstructor = Record<TimestampConstructorKey, Timestamp>;
+
+export type TimestampConstructorKey = "now";
+
+export const Timestamp: TimestampConstructor = {
   get now() {
     return Math.floor(Date.now() / 1000) as Timestamp;
   },
